+++ frontend-api/src/main/kotlin/no/nav/melosys/tjenester/gui/dto/trygdeavgift/TrygdeavgiftsperioderDto.kt
@@ -16,7 +16,7 @@ data class TrygdeavgiftsperioderDto(
             trygdeavgiftsperiode.periodeFra,
             trygdeavgiftsperiode.periodeTil,
             trygdeavgiftsperiode.hentGjeldendeTrygdedekning(),
-            trygdeavgiftsperiode.trygdesats.toDouble(),
+            trygdeavgiftsperiode.trygdesats,
             trygdeavgiftsperiode.trygdeavgiftsbeløpMd.verdi.toInt()
         )
 }
