+++ integrasjon/src/main/kotlin/no/nav/melosys/integrasjon/trygdeavgift/dto/Trygdeavgiftsperiode.kt
@@ -1,9 +1,7 @@
 package no.nav.melosys.integrasjon.trygdeavgift.dto
 
-import java.math.BigDecimal
-
 data class Trygdeavgiftsperiode(
     val periode: DatoPeriode,
-    val sats: BigDecimal,
+    val sats: Double,
     val avgift: PengerDto
 )
